chore: apply comments Signed-off-by: Sonia Sandler <ssandler@redhat.com>
